# Mark GetBundleForChannel as deprecated and trim its response.